Area of Circle

+++ Compoundinterest.py
@@ -1,7 +1,7 @@
-Pr = int(input("enter the Principal amount"))
+Pr = float(input("enter the Principal amount"))
 rate =float(input('Enter the Rate'))
 num= int(input('Number of times interest is compounded per year'))
-time=int(input('no of years'))
+time=float(input('no of years'))
 
 Amount = Pr*(1+rate/num)**(num*time)
 print(Amount)
